requirements.txt
fastapi==0.118.0
uvicorn==0.37.0
pillow==11.3.0
pydantic==2.11.9
python-multipart==0.0.9
